requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.6
crispy-bootstrap5==2026.3
Django==5.2
django-crispy-forms==2.6
django-environ==0.13.0
django-htmx==1.27.0
et_xmlfile==2.0.0
gunicorn==25.3.0
idna==3.11
openpyxl==3.1.5
packaging==26.0
pillow==12.1.1
psycopg2-binary==2.9.11
requests==2.33.0
sqlparse==0.5.5
urllib3==2.6.3
whitenoise==6.12.0
